Fix API break in SQL admin dep.
We updated our clients to all be on the latest available version of the
google.golang.org/api repo, but there's a backwards-incompatible change
to the SQL admin library: it now no longer requires the host to be
passed as an argument when updating a user. The updates to the vendored
libraries have been merged downstream
(terraform-providers/terraform-provider-google#2406 and
terraform-providers/terraform-provider-google-beta#87), but the code is
now broken until we update our usage of the vendored library.

This fixes our usage to make the code compile again.